fix bug of transfers not taking decimals into account fix missing template after wallet transfer for paying lesson

+++ app/interactors/mango/pay_from_wallet.rb
@@ -1,7 +1,7 @@
 module Mango
   class PayFromWallet < BaseInteraction
     object :user, class: User
-    integer :amount
+    float :amount
     string :wallet, default: 'transaction'
     validates :wallet, presence: true
